+++ src/test/java/com/cucumber/stepdefinitions/MyStepdefs.java
@@ -1,9 +1,10 @@
 package com.cucumber.stepdefinitions;
 
 import com.cucumber.listener.Reporter;
+import cucumber.api.DataTable;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
-import org.junit.Assert;
+import cucumber.api.java.en.When;
 
 import java.io.IOException;
 
@@ -13,8 +14,8 @@ public class MyStepdefs {
         throws IOException {
         Reporter.addStepLog("My test addStepLog message");
         Reporter.addScenarioLog("This is scenario log");
-//        Reporter.addScreenCaptureFromPath(
-//            "/Users/vimalrajselvam/Downloads/best-resume-template-2016-3.jpg", "My title");
+        //        Reporter.addScreenCaptureFromPath(
+        //            "/Users/vimalrajselvam/Downloads/best-resume-template-2016-3.jpg", "My title");
     }
 
     @Given("I have (\\d+) cukes in my bellies") public void I_have_cukes_in_my_bellies(int cukes) {
@@ -23,5 +24,8 @@ public class MyStepdefs {
 
     @Then("^I print$") public void iPrint() throws Throwable {
         //        Assert.assertTrue(false);
+    }
+
+    @When("^I login with credentials$") public void iLoginWithCredentials(DataTable table) {
     }
 }